fix: DBAL connection retries when connection was lost
Try and retry an SQL query in case of connection timeouts. This avoids
multiple PING requests in rapid succession. Since all query performed
are meant to be atomic anyway, there should be no lost data and no data
duplication.
